requirements.txt
astroid==2.4.2
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.3
cycler==0.11.0
distlib==0.3.3
Faker==15.0.0
filelock==3.4.0
Flask==2.0.2
fonttools==4.37.0
idna==3.3
isort==5.6.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.4.4
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
matplotlib==3.5.3
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.29
numpy==1.23.2
packaging==21.3
panda==0.3.1
pandas==1.4.3
Pillow==9.2.0
platformdirs==2.4.0
protobuf==4.21.2
pycryptodome==3.13.0
pylint==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
regex==2022.8.17
requests==2.26.0
six==1.15.0
soupsieve==2.3.2.post1
tabulate==0.8.9
tk==0.1.0
toml==0.10.1
urllib3==1.26.7
virtualenv==20.10.0
Werkzeug==2.0.2
wordcloud==1.8.2.2
wrapt==1.12.1
